fixed warning c4244

+++ games/tic_tac_toe.c
@@ -36,7 +36,8 @@ static char game_table[9];
  * @note No checks are included for program execution failures!
  */
 int main()
-{   srand(time(NULL));
+{   
+    srand( (unsigned int)time(NULL));
     int l = 0;
     do
     {
